+++ src/fintrist_ds/__init__.py
@@ -2,7 +2,6 @@
 Import all of the processing functions.
 """
 from .engine import *
-from .catalog import CATALOG
 from .manage import *
 from .services import *
 from .viz import *